src/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
sentence-transformers==3.3.1
torch==2.5.1
psycopg2-binary
httpx
pydantic
